- Minor cleanup
