+++ spa/src/common/LoggedinInfo.tsx
@@ -4,15 +4,14 @@ import { popup } from "../utils";
 import analyticsClient from "../analytics";
 import OAuthManager from "../services/oauth-manager";
 
-
-export const LoggedInContent = styled.div`
+const LoggedInContent = styled.div`
 	display: flex;
 	align-items: center;
 	justify-content: center;
 	margin: 0 auto;
 `;
 
-export const LoggedinInfo = ({ username, logout }: { username: string; logout: () => void;}) => {
+const LoggedinInfo = ({ username, logout }: { username: string; logout: () => void;}) => {
 	const clicked = () => {
 		// Opens the popup for logging out of GitHub
 		popup("https://github.com/logout");
@@ -28,3 +27,5 @@ export const LoggedinInfo = ({ username, logout }: { username: string; logout: (
 		<Button style={{ paddingLeft: 0 }} appearance="link" onClick={clicked}>Change GitHub login</Button>
 	</LoggedInContent>;
 };
+
+export default LoggedinInfo;
